Fixed up Assets
Only need to fix the RVTS to read height and color.
Also need to take a look at the Tree trunk texture and see why its broke. Will no convert to a Virtual Texture as it says its already an Virtual Texture. Its also scaled down very badly.
also the exponetial height fog doesn't seem to be working the same as in our previous project.
I'm also not sure of all the project settings we have enabled in the last project so a double check of that would eb great!